requirements.txt
Django==1.5.1
MySQL-python==1.2.4c1
PyYAML==3.11
South==0.7.6
bitlyapi==0.1.1
coverage==3.7.1
django-braces==1.4.0
django-crispy-forms-1.4.0
django-nose==1.3
django-pagination==1.0.7
django-social-auth==0.7.9
feedparser==5.1.2
html2text==3.200.3
httplib2==0.9
nltk==2.0.4
nose==1.3.4
numpy==1.6.2
oauth2==1.5.211
python-openid==2.2.5
requests==0.14.2
six==1.8.0
tweepy==2.0
virtualenv==1.8.2
wsgiref==0.1.2
